+++ tests/test_plot.py
@@ -1,4 +1,4 @@
-from typing import Optional, cast
+from typing import Literal, Optional, Union, cast
 
 import matplotlib.collections as mcoll
 import matplotlib.pyplot as plt
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 from numpy.typing import NDArray
 
-from bbstat.evaluate import BootstrapResult
+from bbstat.evaluate import BootstrapDistribution
 from bbstat.plot import plot
 
 
@@ -16,36 +16,49 @@ def estimates() -> NDArray[np.floating]:
 
 
 @pytest.fixture(scope="module")
-def bootstrap_result(estimates) -> BootstrapResult:
-    return BootstrapResult(estimates=estimates, level=0.95)
+def bootstrap_distribution(estimates) -> BootstrapDistribution:
+    return BootstrapDistribution(estimates=estimates)
 
 
-def test_plot_returns_axes(bootstrap_result: BootstrapResult) -> None:
-    ax = plot(bootstrap_result)
+@pytest.mark.parametrize(
+    "precision",
+    [
+        pytest.param(None),
+        pytest.param(1),
+        pytest.param("auto"),
+    ],
+)
+def test_plot_returns_axes(
+    bootstrap_distribution: BootstrapDistribution,
+    precision: Optional[Union[int, Literal["auto"]]],
+) -> None:
+    ax = plot(bootstrap_distribution, 0.87, precision=precision)
     assert isinstance(ax, plt.Axes)
 
 
 @pytest.mark.parametrize(
     "level, expected_title",
     [
-        pytest.param(None, "Bayesian bootstrap  •  101 resamples, 95% CI"),
+        pytest.param(0.95, "Bayesian bootstrap  •  101 resamples, 95% CI"),
         pytest.param(0.99, "Bayesian bootstrap  •  101 resamples, 99% CI"),
     ],
 )
 def test_plot_respects_level_in_title(
-    bootstrap_result: BootstrapResult,
-    level: Optional[float],
+    bootstrap_distribution: BootstrapDistribution,
+    level: float,
     expected_title: str,
 ) -> None:
-    ax = plot(bootstrap_result, level=level)
+    ax = plot(bootstrap_distribution, level=level)
     actual_title = ax.get_title()
     assert isinstance(actual_title, str)
     assert actual_title == expected_title
 
 
-def test_plot_adds_three_lines_and_one_fill(bootstrap_result: BootstrapResult) -> None:
+def test_plot_adds_three_lines_and_one_fill(
+    bootstrap_distribution: BootstrapDistribution,
+) -> None:
     fig, ax = plt.subplots()
-    _ = plot(bootstrap_result, ax=ax)
+    _ = plot(bootstrap_distribution, 0.87, ax=ax)
     assert len(ax.lines) == 3
     assert len([c for c in ax.collections if isinstance(c, mcoll.PolyCollection)]) == 1
 
@@ -58,12 +71,12 @@ def test_plot_adds_three_lines_and_one_fill(bootstrap_result: BootstrapResult) -
     ],
 )
 def test_plot_labels_match(
-    bootstrap_result: BootstrapResult,
+    bootstrap_distribution: BootstrapDistribution,
     label: Optional[str],
     expected_label: str,
 ) -> None:
     label = "my_stat"
-    ax = plot(bootstrap_result, label=label)
+    ax = plot(bootstrap_distribution, 0.87, label=label)
     actual_label = ax.lines[0].get_label()
     assert isinstance(label, str)
     cast(str, actual_label).startswith(expected_label)